refactor(blorktools): file handlers
